backing up for the day, make a few plans just take devices
