+++ src/ibex_bluesky_core/plan_stubs/__init__.py
@@ -6,6 +6,8 @@ from typing import Callable, ParamSpec, TypeVar, cast
 import bluesky.plan_stubs as bps
 from bluesky.utils import Msg
 
+from ibex_bluesky_core.devices.simpledae import SimpleDae
+
 P = ParamSpec("P")
 T = TypeVar("T")
 
@@ -78,3 +80,20 @@ def call_qt_aware(
 
     yield from bps.clear_checkpoint()
     return cast(T, (yield Msg(CALL_QT_AWARE_MSG_KEY, func, *args, **kwargs)))
+
+
+def set_num_periods(dae: SimpleDae, nperiods: int) -> Generator[Msg, None, None]:
+    """Set the number of software periods for a DAE.
+
+    Args:
+        dae (SimpleDae): DAE object.
+        nperiods (int): number of software periods to set.
+
+    """
+    yield from bps.mv(dae.number_of_periods, nperiods)  # type: ignore
+    actual = yield from bps.rd(dae.number_of_periods)
+    if actual != nperiods:
+        raise ValueError(
+            f"Could not set {nperiods} periods on DAE (probably requesting too many points, "
+            f"or already running)"
+        )
