+++ src/ibex_bluesky_core/utils/__init__.py
@@ -1,0 +1,16 @@
+"""Utilities for plans which are not plan stubs."""
+
+
+def centred_pixel(centre: int, pixel_range: int) -> list[int]:
+    """Given a centre and range, return a contiguous range of pixels around the centre, inclusive.
+
+    ie. a centre of 50 with a range of 3 will give [47, 48, 49, 50, 51, 52, 53]
+
+    Args:
+          centre (int): The centre pixel number.
+          pixel_range (int): The range of pixels either side to surround the centre.
+
+    Returns a list of pixel numbers.
+
+    """
+    return [s for s in range(centre - pixel_range, centre + pixel_range + 1)]
